requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
boto3==1.20.34
botocore==1.23.34
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
jmespath==0.10.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
